refactor: binary parser
