--- packages/apidom-reference/src/util/btoa/index-node.ts
@@ -1,8 +1,0 @@
-import { Buffer } from 'node:buffer';
-
-const btoa = (val: string) => {
-  const buffer = Buffer.from(val.toString(), 'binary');
-  return buffer.toString('base64');
-};
-
-export default btoa;
