+++ packages/apidom-reference/test/parse/parsers/apidom-reference-parser-binary/index.ts
@@ -3,7 +3,7 @@ import { assert } from 'chai';
 import { isParseResultElement, isStringElement } from '@swagger-api/apidom-core';
 
 import File from '../../../../src/util/File';
-import BinaryParser from '../../../../src/parse/parsers/apidom-reference-parser-binary';
+import BinaryParser from '../../../../src/parse/parsers/apidom-reference-parser-binary/index-node';
 import { ParserError } from '../../../../src/util/errors';
 
 describe('parsers', function () {
